requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1 